% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfunct.R
\name{costfunct}
\alias{costfunct}
\title{Cost function implementing C2 in the binary segmentation article}
\usage{
costfunct(data.vec)
}
\arguments{
\item{data.vec}{Input data vector}
}
\value{
The sum of squares of the data vector
}
\description{
Cost function implementing C2 in the binary segmentation article
}
\examples{
test.vec = c(1,2,3,5,7,9,11)
sum.sq = costfunct(test.vec)
}
