% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.r
\name{PCA}
\alias{PCA}
\title{Principal Components Analysis}
\usage{
PCA(data.mat)
}
\arguments{
\item{data.mat}{Input data matrix}
}
\value{
Object of class tomPCA with rotation and x data
}
\description{
Principal Components Analysis
}
\examples{
test.matrix = matrix(
   c(2, 4, 3, 1, 5, 7),
   nrow=3,
   ncol=2)

pca.obj = PCA(test.matrix)
}
